tensorflow_scripts/bounding_box_hub/requirements.txt
tensorflow==2.3.0
requests==2.25.1
pydocstyle==5.1.1
flake8==3.9.0
pylint==2.7.2
pytest==6.2.3
Pillow==7.2.0
numpy==1.18.5
opencv-python
datature-hub==0.2.0